Factor out common backend tests
