+++ tests/test_backends/test_theano.py
@@ -6,43 +6,26 @@ from numpy import random as rnd, testing as np_testing
 
 from pymanopt.function import Theano
 
+from . import _backend_tests
 
-class TestArity(unittest.TestCase):
-    def test_unary_parameter(self):
-        """
-        Test cost function, gradient and Hessian for a simple unary function.
-        """
+
+class TestUnaryFunction(_backend_tests.TestUnaryFunction):
+    def setUp(self):
+        super().setUp()
+
         x = T.vector()
 
         @Theano(x)
         def cost(x):
             return T.sum(x ** 2)
 
-        n = 10
-        x = rnd.randn(n)
+        self.cost = cost
 
-        # Test whether cost function accepts single argument.
-        self.assertAlmostEqual(np.sum(x ** 2), cost(x))
 
-        # Test whether gradient accepts single argument.
-        egrad = cost.compute_gradient()
-        np_testing.assert_allclose(2 * x, egrad(x))
+class TestNaryFunction(_backend_tests.TestNaryFunction):
+    def setUp(self):
+        super().setUp()
 
-        # Test the Hessian.
-        u = rnd.randn(n)
-
-        # Test whether Hessian accepts two regular arguments.
-        ehess = cost.compute_hessian()
-        # Test whether Hessian-vector product is correct.
-        np_testing.assert_allclose(2 * u, ehess(x, u))
-
-    def test_unary_nary_parameter(self):
-        """
-        Test cost function, gradient and Hessian when multiple arguments are
-        grouped into a single element. This situation arises e.g. when
-        optimizing over the FixedRankEmbedded manifold where points on the
-        manifold are represented as a 3-tuple making up a truncated SVD.
-        """
         x = T.vector()
         y = T.vector()
 
@@ -50,49 +33,13 @@ class TestArity(unittest.TestCase):
         def cost(x, y):
             return T.dot(x, y)
 
-        n = 10
-        x = rnd.randn(n)
-        y = rnd.randn(n)
+        self.cost = cost
 
-        # The argument signature of the cost function implies we are NOT on the
-        # product manifold so solvers would call the wrapped cost function with
-        # one argument, in this case a tuple of vectors.
-        self.assertAlmostEqual(np.dot(x, y), cost((x, y)))
 
-        egrad = cost.compute_gradient()
-        g = egrad((x, y))
-        # Since we treat the tuple (x, y) as one argument, we expect the result
-        # of a call to the gradient function to be a tuple with two elements.
-        self.assertIsInstance(g, (list, tuple))
-        self.assertEqual(len(g), 2)
-        for gi in g:
-            self.assertIsInstance(gi, np.ndarray)
-        g_x, g_y = g
-        np_testing.assert_allclose(g_x, y)
-        np_testing.assert_allclose(g_y, x)
+class TestNaryParameterGrouping(_backend_tests.TestNaryParameterGrouping):
+    def setUp(self):
+        super().setUp()
 
-        # Test the Hessian-vector product.
-        u = rnd.randn(n)
-        v = rnd.randn(n)
-
-        ehess = cost.compute_hessian()
-        h = ehess((x, y), (u, v))
-        self.assertIsInstance(h, (list, tuple))
-        self.assertEqual(len(h), 2)
-        for hi in h:
-            self.assertIsInstance(hi, np.ndarray)
-
-        # Test whether the Hessian-vector product is correct.
-        h_x, h_y = h
-        np_testing.assert_allclose(h_x, v)
-        np_testing.assert_allclose(h_y, u)
-
-    def test_nary_parameter_grouping(self):
-        """
-        Test cost function, gradient and Hessian for a complex cost function
-        one would define on product manifolds where one of the underlying
-        manifolds represents points as a tuple of numpy.ndarrays.
-        """
         x = T.vector()
         y = T.vector()
         z = T.vector()
@@ -101,50 +48,7 @@ class TestArity(unittest.TestCase):
         def cost(x, y, z):
             return T.sum(x ** 2 + y + z ** 3)
 
-        n = 10
-        x, y, z = [rnd.randn(n) for _ in range(3)]
-
-        # The signature of the cost function now implies that we are on the
-        # product manifold, so we mimic the behavior of solvers by calling the
-        # cost function with a single argument: a tuple containing a tuple (x,
-        # y) and a single vector z.
-        self.assertAlmostEqual(np.sum(x ** 2 + y + z ** 3), cost(((x, y), z)))
-
-        egrad = cost.compute_gradient()
-        g = egrad(((x, y), z))
-        # We defined the cost function signature to treat the first two
-        # arguments as one parameter, so a call to the gradient must produce
-        # two elements.
-        self.assertIsInstance(g, (list, tuple))
-        self.assertEqual(len(g), 2)
-        g_xy, g_z = g
-        self.assertIsInstance(g_xy, (list, tuple))
-        self.assertEqual(len(g_xy), 2)
-        self.assertIsInstance(g_z, np.ndarray)
-
-        # Verify correctness of the gradient.
-        np_testing.assert_allclose(g_xy[0], 2 * x)
-        np_testing.assert_allclose(g_xy[1], 1)
-        np_testing.assert_allclose(g_z, 3 * z ** 2)
-
-        # Test the Hessian.
-        u, v, w = [rnd.randn(n) for _ in range(3)]
-
-        ehess = cost.compute_hessian()
-        h = ehess(((x, y), z), ((u, v), w))
-
-        # Test the type composition of the return value.
-        self.assertIsInstance(h, (list, tuple))
-        self.assertEqual(len(h), 2)
-        h_xy, h_z = h
-        self.assertIsInstance(h_xy, (list, tuple))
-        self.assertEqual(len(h_xy), 2)
-        self.assertIsInstance(h_z, np.ndarray)
-
-        # Test whether the Hessian-vector product is correct.
-        np_testing.assert_allclose(h_xy[0], 2 * u)
-        np_testing.assert_allclose(h_xy[1], 0)
-        np_testing.assert_allclose(h_z, 6 * z * w)
+        self.cost = cost
 
 
 class TestVector(unittest.TestCase):
